refactor(#3109): consolidate 10 more test compileAndRun helpers (slice 3)
Migrate the remaining code-identical (modulo comments) local compileAndRun
clusters to shared helpers in tests/helpers/compile.ts:

- compileAndRunIRVariant: 6 tests/equivalence/ir-slice* IR-vs-legacy files
  (identical local helper + identical ENV_STUB console stubs).
- compileAndRunHoistExports: issue-298, var-hoisting (binary-length guard).
- compileAndRunVecSetExports: issue-1441, issue-1057 (setExports/__vec_len).

Each file deletes its local helper (+ orphaned compile/buildImports/ENV_STUB)
and imports the shared function via the { X as compileAndRun } alias, so call
sites are unchanged. Byte-for-byte behavior-preserving; zero src/ changes.

Parity: vitest run of the 10 files = 52 tests all pass; tsc/biome/prettier clean.
Count: 132 -> 52 removed; 80 local definitions remain (issue stays ready until
>=100 removed).

✓
