+++ tests/issue-1441.test.ts
@@ -1,20 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.ts";
-import { buildImports } from "../src/runtime.ts";
-
-async function compileAndRun(source: string): Promise<any> {
-  const result = await compile(source, { fileName: "test.ts" });
-  if (!result.success) {
-    throw new Error(`Compile error: ${result.errors?.[0]?.message}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const mod = new WebAssembly.Module(result.binary);
-  const instance = new WebAssembly.Instance(mod, imports);
-  // Wire wasmExports so the runtime can reach `__vec_len` for the
-  // `constructor === Array` lookup on vec wrapper structs.
-  imports.setExports?.(instance.exports as Record<string, Function>);
-  return (instance.exports as any).test();
-}
+import { compileAndRunVecSetExports as compileAndRun } from "./helpers/compile.js";
 
 describe("#1441 — String.prototype.split: Array shape + wrapper receivers + limit", () => {
   describe("Array result shape (.constructor === Array)", () => {
