+++ tests/issue-298.test.ts
@@ -1,17 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.js";
-import { buildImports } from "../src/runtime.js";
-
-async function compileAndRun(source: string) {
-  const result = await compile(source, { fileName: "test.ts" });
-  // Don't check TS type errors — var hoisting triggers "used before being assigned"
-  if (!result.binary || result.binary.length === 0) {
-    throw new Error(`Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, imports);
-  return instance.exports as Record<string, Function>;
-}
+import { compileAndRunHoistExports as compileAndRun } from "./helpers/compile.js";
 
 describe("issue-298: function statement edge cases", () => {
   it("nested function with side effects on captured var (S13.2.1_A9.1_T1 pattern)", async () => {
