+++ tests/var-hoisting.test.ts
@@ -1,18 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { compile } from "../src/index.js";
-import { buildImports } from "../src/runtime.js";
-
-async function compileAndRun(source: string) {
-  const result = await compile(source, { fileName: "test.ts" });
-  // Don't check TS type errors — var hoisting triggers "used before being assigned"
-  // which is a TS diagnostic, not a real codegen failure
-  if (!result.binary || result.binary.length === 0) {
-    throw new Error(`Compile failed:\n${result.errors.map((e) => `  L${e.line}: ${e.message}`).join("\n")}`);
-  }
-  const imports = buildImports(result.imports, undefined, result.stringPool);
-  const { instance } = await WebAssembly.instantiate(result.binary, imports);
-  return instance.exports as Record<string, Function>;
-}
+import { compileAndRunHoistExports as compileAndRun } from "./helpers/compile.js";
 
 describe("var hoisting", () => {
   it("var in for-loop initializer accessible after loop", async () => {
